+++ src/graphics/fonts/trueTypeFont.cs
@@ -191,7 +191,7 @@ namespace Graphics
          Renderer.device.drawIndexed(PrimitiveType.TriangleStrip, myIbo.count, 0, DrawElementsType.UnsignedShort);
       }
 
-      public override int width(String txt)
+      public override float width(String txt)
       {
          int size = 0;
          for (int i = 0; i < txt.Length; i++)
@@ -203,7 +203,7 @@ namespace Graphics
          return size;
       }
 
-      public override int height(String txt)
+      public override float height(String txt)
       {
          return mySize;
       }
@@ -219,7 +219,7 @@ namespace Graphics
          if (xCounter + glyph.bitmap.width > texture.width)
          {
             xCounter = 0;
-            yCounter += mySize;
+            yCounter += (int)mySize;
          }
 
          if (yCounter + glyph.bitmap.rows > texture.height)
